requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
anyio==3.3.4
asgiref==3.4.1
astunparse==1.6.3
async-cse==0.3.0
async-generator==1.10
async-timeout==4.0.1
asyncpg==0.25.0
attrs==21.2.0
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.10.0
braceexpand==0.1.7
bs4==0.0.1
CacheControl==0.12.10
cachy==0.3.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
cleo==0.8.1
click==8.0.3
clikit==0.6.2
colorama==0.4.4
colorthief==0.2.1
crashtest==0.3.1
cryptography==35.0.0
cycler==0.11.0
discord.py @ git+https://github.com/iDevision/enhanced-discord.py@2c7d3c0f89d781fd420db291d0938c2e9422fea6
distlib==0.3.3
dnspython==2.1.0
emoji==1.6.1
fastapi==0.70.0
filelock==3.4.0
Flask==2.0.2
frozenlist==1.2.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.11.1
html5lib==1.1
httpcore==0.9.1
httptools==0.2.0
httpx==0.13.3
humanize==3.12.0
hyperframe==5.2.0
idna==2.10
import-expression==1.1.4
itsdangerous==2.0.1
Jinja2==3.0.3
jishaku==2.3.1
keyring==21.8.0
kiwisolver==1.3.2
lockfile==0.12.2
MarkupSafe==2.0.1
motor==2.5.1
msgpack==1.0.3
multidict==5.2.0
numpy==1.21.4
outcome==1.1.0
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
Pillow==8.4.0
pkginfo==1.8.1
platformdirs==2.4.0
poetry==1.1.11
poetry-core==1.0.7
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.8.2
pylev==1.4.0
pymongo==3.12.1
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
python-vlc==3.0.12118
pywin32-ctypes==0.2.0
PyYAML==6.0
ranimegen==0.1.0
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
selenium==4.0.0
shellingham==1.4.0
simpcalci==0.0.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
starlette==0.16.0
tomlkit==0.7.2
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
virtualenv==20.10.0
waifuim.py==2.4.4
watchgod==0.7
webencodings==0.5.1
websockets==10.1
Werkzeug==2.0.2
wikipedia==1.4.0
wsproto==1.0.0
yarl==1.7.2
youtube-dl==2021.6.6
